+++ client/src/test/java/com/streamr/client/crypto/KeysRsaTest.java
@@ -4,21 +4,20 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.streamr.client.utils.EncryptionUtil;
 import java.security.KeyPair;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class KeysTest {
+class KeysRsaTest {
   @Test
   void validatePublicKeyValidFormat() {
     KeyPair keyPair = EncryptionUtil.generateKeyPair();
     String publicKey = EncryptionUtil.exportPublicKeyAsPemString(keyPair.getPublic());
-    Keys.validatePublicKey(publicKey);
+    KeysRsa.validatePublicKey(publicKey);
   }
 
   @Test
   void validatePublicKeyThrowsOnWrongFormat() {
     try {
-      Keys.validatePublicKey("wrong-format");
+      KeysRsa.validatePublicKey("wrong-format");
     } catch (IllegalArgumentException e) {
       assertEquals("Must be a valid RSA public key in the PEM format.", e.getMessage());
     }
@@ -28,13 +27,13 @@ class KeysTest {
   void validatePrivateKeyValidFormat() {
     KeyPair keyPair = EncryptionUtil.generateKeyPair();
     String privateKey = EncryptionUtil.exportPrivateKeyAsPemString(keyPair.getPrivate());
-    Keys.validatePrivateKey(privateKey);
+    KeysRsa.validatePrivateKey(privateKey);
   }
 
   @Test
   void validatePrivateKeyThrowsOnWrongFormat() {
     try {
-      Keys.validatePrivateKey("wrong-format");
+      KeysRsa.validatePrivateKey("wrong-format");
     } catch (IllegalArgumentException e) {
       assertEquals("Must be a valid RSA private key in the PEM format.", e.getMessage());
     }
